Retry the database check if it fails
If the app is running in a container and the database check fails,
the container will exit. Production deployments may automatically
try to restart the container; however, that will also fail unless
the database has been initialized. The easiest way to initialize
the database may be to `exec` the init command in the container,
though, which will not work unless the container is running.

This keeps the container running and re-checking the database
every 10 seconds, which should allow the container to be used for
initialization without bombarding the database.