Redirect to a raw txt retrieval from the Web UI
Currently we redirect to a Web render;
however, if libmagic guesses a type that the browser cannot render,
the browser will try to download the paste,
which makes it difficult to get a link to the paste.

This makes it so the Web UI redirects to a raw txt render.
The user can make it prettier my manually adjusting the URL if they choose.